+++ NEWS.md
@@ -1,5 +1,7 @@
 # httr2 (development version)
 
+* `resp_body_json/html/xml()` now work for cached responses (@mgirlich, #277).
+
 * `with_mock()` and `local_mock()` now correctly trigger errors when the
   mocked response represents an HTTP failure (#252).